edits
Okay SO i replaced the pictures of the basses because they were too fuzzy. Replaced it with a slightly less fuzzy picture from Pops musics instagram. Added a picture to the lessons page. Started adding links to the store page and making the filters work and everything. Added a picture to the repairs page